+++ packages/sdk-router/src/module/index.ts
@@ -1,2 +1,6 @@
+export * from './query'
+
 export * from './synapseModule'
 export * from './synapseModuleSet'
+
+export * from './types'
